Refactor database operations and UI logic for transactions, wallets, and categories to support soft deletion, synchronization, and improved data consistency.
### Data & Core
- **Database Architecture**: Incremented `DATABASE_VERSION` to 5. Updated `TransactionEntry` to allow nullable `COLUMN_CATEGORY_ID` for better referential integrity during category deletion.
- **DAO Enhancements**:
    - Centralized transaction join queries in `TransactionDao` to include category and wallet details.
    - Implemented soft delete (updating `deleted_at` and `updated_at`) in `CategoryDao`, `WalletDao`, and `TransactionDao`.
    - Added database indexes for sync support, wallets, and budgets to improve query performance.
    - Updated `CategoryDao` to handle system vs. user categories and manage `created_at`/`updated_at` timestamps.
- **Repositories**:
    - Refactored `TransactionRepository` to use `long` for monetary values instead of `double` to prevent precision issues.
    - Optimized transaction filtering by delegating date range logic directly to the database layer.

### UI & ViewModels
- **Transaction Management**:
    - Separated `WalletPickerAdapter` and `CategoryPickerAdapter` into independent files.
    - Updated `TransactionViewModel` to use `Long` for balance/expense/income totals and improved empty state handling.
- **Wallet & Category Logic**:
    - Enhanced `WalletViewModel` with a `loadWalletById` method and standardized `SharedPreferences` keys for active wallets across the app.
    - Improved `EditWalletFragment` to load data independently of the main wallet list, preventing issues during screen rotations.
    - Fixed category management logic to consistently reload both system and user-defined categories.
- **Budgeting**:
    - Introduced `OperationResult` with `Action` types (CREATE, UPDATE, DELETE) and a "handled" state to prevent duplicate Toast notifications.
- **Home Screen**: Refactored `HomeViewModel` to fetch real transaction data for the "Top Expenses" view instead of using mock data.

### Formatting & Consistency
- Standardized currency formatting to use integer representation (`%,d đ`) across all transaction lists and summaries.
- Improved `Fragment` lifecycle handling in `WalletFragment` to ensure the wallet list refreshes on resume.

+++ app/src/main/java/com/ptithcm/quanlichitieu/data/model/TransactionGroup.java
@@ -6,10 +6,10 @@ public class TransactionGroup {
 
     private final String dayOfWeek;
     private final String date;
-    private final double dayTotal;
+    private final long dayTotal;
     private final List<Transaction> transactions;
 
-    public TransactionGroup(String dayOfWeek, String date, double dayTotal, List<Transaction> transactions) {
+    public TransactionGroup(String dayOfWeek, String date, long dayTotal, List<Transaction> transactions) {
         this.dayOfWeek = dayOfWeek;
         this.date = date;
         this.dayTotal = dayTotal;
@@ -18,6 +18,6 @@ public class TransactionGroup {
 
     public String getDayOfWeek() { return dayOfWeek; }
     public String getDate() { return date; }
-    public double getDayTotal() { return dayTotal; }
+    public long getDayTotal() { return dayTotal; }
     public List<Transaction> getTransactions() { return transactions; }
 }
